(fix) Fixed the default login location coming in the every term search
