+++ packages/apps/esm-login-app/src/login.resource.ts
@@ -1,17 +1,4 @@
-import { useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-import useSwrInfinite from 'swr/infinite';
-import useSwrImmutable from 'swr/immutable';
-import {
-  FetchResponse,
-  fhirBaseUrl,
-  openmrsFetch,
-  refetchCurrentUser,
-  Session,
-  showNotification,
-  useDebounce,
-} from '@openmrs/esm-framework';
-import { LocationEntry, LocationResponse } from './types';
+import { openmrsFetch, refetchCurrentUser, Session } from '@openmrs/esm-framework';
 
 export async function performLogin(username: string, password: string): Promise<{ data: Session }> {
   const abortController = new AbortController();
@@ -27,114 +14,4 @@ export async function performLogin(username: string, password: string): Promise<
     refetchCurrentUser();
     return res;
   });
-}
-
-interface LoginLocationData {
-  locations: Array<LocationEntry>;
-  isLoading: boolean;
-  totalResults: number;
-  hasMore: boolean;
-  loadingNewData: boolean;
-  setPage: (size: number | ((_size: number) => number)) => Promise<FetchResponse<LocationResponse>[]>;
-}
-
-export function useLoginLocations(
-  useLoginLocationTag: boolean,
-  count: number = 0,
-  searchQuery: string = '',
-): LoginLocationData {
-  const { t } = useTranslation();
-  const debouncedSearchQuery = useDebounce(searchQuery);
-  function constructUrl(page: number, prevPageData: FetchResponse<LocationResponse>) {
-    if (prevPageData) {
-      const nextLink = prevPageData.data?.link?.find((link) => link.relation === 'next');
-
-      if (!nextLink) {
-        return null;
-      }
-
-      const nextUrl = new URL(nextLink.url);
-      // default for production
-      if (nextUrl.origin === window.location.origin) {
-        return nextLink.url;
-      }
-
-      // in development, the request should be funnelled through the local proxy
-      return new URL(
-        `${nextUrl.pathname}${nextUrl.search ? `?${nextUrl.search}` : ''}`,
-        window.location.origin,
-      ).toString();
-    }
-
-    let url = `${fhirBaseUrl}/Location?`;
-    let urlSearchParameters = new URLSearchParams();
-    urlSearchParameters.append('_summary', 'data');
-
-    if (count) {
-      urlSearchParameters.append('_count', '' + count);
-    }
-
-    if (page) {
-      urlSearchParameters.append('_getpagesoffset', '' + page * count);
-    }
-
-    if (useLoginLocationTag) {
-      urlSearchParameters.append('_tag', 'Login Location');
-    }
-
-    if (typeof debouncedSearchQuery === 'string' && debouncedSearchQuery != '') {
-      urlSearchParameters.append('name:contains', debouncedSearchQuery);
-    }
-
-    return url + urlSearchParameters.toString();
-  }
-
-  const { data, isLoading, isValidating, setSize, error } = useSwrInfinite<FetchResponse<LocationResponse>, Error>(
-    constructUrl,
-    openmrsFetch,
-  );
-
-  if (error) {
-    showNotification({
-      title: t('errorLoadingLoginLocations', 'Error loading login locations'),
-      kind: 'error',
-      critical: true,
-      description: error?.message,
-    });
-  }
-
-  const memoizedLocations = useMemo(() => {
-    return {
-      locations: data?.length ? data?.flatMap((entries) => entries?.data?.entry ?? []) : null,
-      isLoading,
-      totalResults: data?.[0]?.data?.total ?? null,
-      hasMore: data?.length ? data?.[data.length - 1]?.data?.link?.some((link) => link.relation === 'next') : false,
-      loadingNewData: isValidating,
-      setPage: setSize,
-    };
-  }, [isLoading, data, isValidating, setSize]);
-
-  return memoizedLocations;
-}
-
-export function useValidateLocationUuid(locationUuid: string) {
-  const url = locationUuid ? `/ws/fhir2/R4/Location?_id=${locationUuid}` : null;
-  const { data, error, isLoading } = useSwrImmutable<FetchResponse<LocationResponse>>(url, openmrsFetch, {
-    shouldRetryOnError(err) {
-      if (err?.response?.status) {
-        return err.response.status >= 500;
-      }
-      return false;
-    },
-  });
-  const results = useMemo(
-    () => ({
-      isLocationValid: data?.ok && data?.data?.total > 0,
-      defaultLocation: data?.data?.entry ?? [],
-      error,
-      isLoading,
-    }),
-    [data, isLoading, error],
-  );
-  return results;
 }
